#25 Moved the sup category code into its own file
